requirements.txt
airbyte==0.33.1
airbyte-api==0.53.0
airbyte-cdk==7.5.1
airbyte_protocol_models_dataclasses==0.17.1
airbyte_protocol_models_pdv2==0.13.1
annotated-types==0.7.0
anyascii==0.3.3
anyio==4.11.0
asn1crypto==1.5.1
asttokens==3.0.1
attributes-doc==0.4.0
attrs==25.4.0
Authlib==1.6.5
backoff==2.2.1
backports.tarfile==1.2.0
beartype==0.22.5
boltons==25.0.0
boto3==1.40.76
botocore==1.40.76
bracex==2.6
cachetools==6.2.2
cattrs==25.3.0
certifi==2025.11.12
cffi==1.17.1
charset-normalizer==3.4.4
click==8.3.1
comm==0.2.3
cryptography==44.0.3
cyclopts==4.2.4
dataclasses-json==0.6.7
dateparser==1.2.2
debugpy==1.8.17
decorator==5.2.1
diskcache==5.6.3
dnspython==2.8.0
docstring_parser==0.17.0
docutils==0.22.3
dpath==2.2.0
duckdb==1.4.2
duckdb_engine==0.13.6
dunamai==1.25.0
email-validator==2.3.0
et_xmlfile==2.0.0
exceptiongroup==1.3.0
executing==2.2.1
fastmcp==2.13.1
filelock==3.20.0
genson==1.3.0
google-api-core==2.28.1
google-auth==2.43.0
google-cloud-bigquery==3.38.0
google-cloud-bigquery-storage==2.34.0
google-cloud-core==2.5.0
google-cloud-secret-manager==2.25.0
google-crc32c==1.7.1
google-resumable-media==2.8.0
googleapis-common-protos==1.72.0
greenlet==3.2.4
grpc-google-iam-v1==0.14.3
grpcio==1.76.0
grpcio-status==1.76.0
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
httpx-sse==0.4.3
idna==3.11
importlib_metadata==8.7.0
ipykernel==7.1.0
ipython==9.7.0
ipython_pygments_lexers==1.1.1
isodate==0.6.1
jaraco.classes==3.4.0
jaraco.context==6.0.1
jaraco.functools==4.3.0
jedi==0.19.2
jeepney==0.9.0
Jinja2==3.1.6
jmespath==1.0.1
joblib==1.5.2
jsonpath-python==1.0.6
jsonref==1.1.0
jsonschema==4.25.1
jsonschema-path==0.3.4
jsonschema-specifications==2025.9.1
jupyter_client==8.6.3
jupyter_core==5.9.1
keyring==25.7.0
markdown-it-py==4.0.0
MarkupSafe==3.0.3
marshmallow==3.26.1
matplotlib-inline==0.2.1
mcp==1.21.2
mdurl==0.1.2
more-itertools==10.8.0
mypy_extensions==1.1.0
nest-asyncio==1.6.0
nltk==3.9.1
numpy==2.3.5
openapi-pydantic==0.5.1
openpyxl==3.1.5
orjson==3.11.4
overrides==7.7.0
packaging==25.0
pandas==2.2.3
parso==0.8.5
pathable==0.4.4
pathvalidate==3.3.1
pexpect==4.9.0
platformdirs==4.5.0
prompt_toolkit==3.0.52
proto-plus==1.26.1
protobuf==6.33.1
psutil==7.1.3
psycopg==3.2.12
psycopg-binary==3.2.12
psycopg-pool==3.2.7
psycopg2-binary==2.9.11
ptyprocess==0.7.0
pure_eval==0.2.3
py-key-value-aio==0.2.8
py-key-value-shared==0.2.8
pyarrow==21.0.0
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.23
pydantic==2.12.4
pydantic-settings==2.12.0
pydantic_core==2.41.5
Pygments==2.19.2
PyJWT==2.10.1
pyOpenSSL==25.1.0
pyperclip==1.11.0
pyrate-limiter==3.1.1
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
python-multipart==0.0.20
python-ulid==3.1.0
pytz==2024.2
PyYAML==6.0.3
pyzmq==27.1.0
RapidFuzz==3.14.3
referencing==0.36.2
regex==2025.11.3
requests==2.32.5
requests-cache==1.2.1
rich==13.9.4
rich-click==1.9.4
rich-rst==1.3.2
rpds-py==0.29.0
rsa==4.9.1
s3transfer==0.14.0
SecretStorage==3.4.1
serpyco-rs==1.17.1
six==1.17.0
sniffio==1.3.1
snowflake-connector-python==3.18.0
snowflake-sqlalchemy==1.7.7
sortedcontainers==2.4.0
SQLAlchemy==2.0.44
sqlalchemy-bigquery==1.12.0
sse-starlette==3.0.3
stack-data==0.6.3
starlette==0.50.0
structlog==24.4.0
tomlkit==0.13.3
tornado==6.5.2
tqdm==4.67.1
traitlets==5.14.3
typing-inspect==0.9.0
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.2
tzlocal==5.3.1
Unidecode==1.4.0
url-normalize==2.2.1
urllib3==2.5.0
uuid7==0.1.0
uv==0.8.24
uvicorn==0.38.0
wcmatch==10.0
wcwidth==0.2.14
websockets==15.0.1
whenever==0.8.10
xmltodict==0.14.2
zipp==3.23.0
